RDRP-671: moved create_period_year to form prep and changed schemas
