+++ config/output_schemas/gb_sas_schema.toml
@@ -3,7 +3,7 @@ old_name = "reference"
 Deduced_Data_Type = "Int64"
 
 [period]
-old_name = "period"
+old_name = "period_year"
 Deduced_Data_Type = "Int64"
 
 [entref]
