+++ src/outputs/short_form.py
@@ -7,7 +7,7 @@ from typing import Callable, Dict, Any
 import src.outputs.map_output_cols as map_o
 from src.staging.validation import load_schema
 from src.imputation.imputation_helpers import fill_sf_zeros
-from src.outputs.outputs_helpers import create_output_df, create_period_year
+from src.outputs.outputs_helpers import create_output_df
 
 
 OutputMainLogger = logging.getLogger(__name__)
@@ -110,9 +110,6 @@ def output_short_form(
     NETWORK_OR_HDFS = config["global"]["network_or_hdfs"]
     paths = config[f"{NETWORK_OR_HDFS}_paths"]
     output_path = paths["output_path"]
-
-    # Create a 'year' column
-    df = create_period_year(df)
 
     # Join foriegn ownership column using ultfoc mapper
     df = map_o.join_fgn_ownership(df, ultfoc_mapper)
