+++ tests/test_outputs/test_short_form.py
@@ -3,32 +3,7 @@ import pandas as pd
 import unittest
 from pandas._testing import assert_frame_equal
 
-from src.outputs.short_form import create_period_year, create_headcount_cols
-
-
-class TestCreateNewCols(unittest.TestCase):
-    """Unittest for create news columns"""
-
-    def data_frame(self):
-        # create sample data
-        data = {"period": ["202201", "202206", "202109", "202112"]}
-
-        df = pd.DataFrame(data)
-        return df
-
-    def test_create_new_cols(self):
-        # Call the create_new_cols funtion
-        df_input = self.data_frame()
-        actual_result = create_period_year(df_input)
-
-        expected_result = pd.DataFrame(
-            {
-                "period": ["202201", "202206", "202109", "202112"],
-                "period_year": [2022, 2022, 2021, 2021],
-            }
-        )
-
-        assert_frame_equal(actual_result, expected_result)
+from src.outputs.short_form import create_headcount_cols
 
 
 class TestCreateHeadcountCols:
